requirements-ci.txt
psycopg2-binary
pytest
pytest-cov
ruff
